完善数据库接口

+++ example/src/main/java/com/liux/android/example/downloader/SingleTaskDownloaderActivity.java
@@ -7,6 +7,7 @@ import android.support.v7.app.AppCompatActivity;
 import com.liux.android.downloader.Downloader;
 import com.liux.android.downloader.InitCallback;
 import com.liux.android.downloader.OnStatusListener;
+import com.liux.android.downloader.UIStatusListener;
 import com.liux.android.downloader.core.Task;
 import com.liux.android.example.R;
 
@@ -35,19 +36,19 @@ public class SingleTaskDownloaderActivity extends AppCompatActivity {
     @Override
     protected void onStart() {
         super.onStart();
-        task.start();
+        if (task != null) task.start();
     }
 
     @Override
     protected void onStop() {
         super.onStop();
-        task.stop();
+        if (task != null) task.stop();
     }
 
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        task.delete();
+        if (task != null) task.delete();
     }
 
     private void onCreateTask() {
@@ -59,12 +60,40 @@ public class SingleTaskDownloaderActivity extends AppCompatActivity {
                 .build();
         task.bindStatusListener(new OnStatusListener() {
             @Override
-            public void onBind() {
-                onUpdate();
+            public void onBind(Task task) {
+                onUpdate(task);
             }
 
             @Override
-            public void onUpdate() {
+            public void onUpdate(Task task) {
+                switch (task.getStatus()) {
+                    case NEW:
+                        break;
+                    case WAIT:
+                        break;
+                    case CONN:
+                        break;
+                    case START:
+                        break;
+                    case STOP:
+                        break;
+                    case ERROR:
+                        break;
+                    case COMPLETE:
+                        break;
+                    case DELETE:
+                        break;
+                }
+            }
+        });
+        task.bindStatusListener(new UIStatusListener() {
+            @Override
+            protected void onUIBind(Task task) {
+                onUIUpdate(task);
+            }
+
+            @Override
+            protected void onUIUpdate(Task task) {
                 switch (task.getStatus()) {
                     case NEW:
                         break;
